chat_bot/requirements.txt
datetime==4.0.1
numpy==1.22.4
pandas==1.5.3
plotly==5.13.1
sklearn-pandas==2.2.0
stocknews==0.9.11
streamlit==1.24.1
tensorflow==2.12.0
yfinance==0.2.38
crewai==0.11.0
duckduckgo-search
bs4
undetected-chromedriver
langchain
langchain-google-vertexai
websocket-client
pandas
ta-lib
yfinance
plotly
streamlit
textblob
requests
newspaper3k